+++ NEWS.md
@@ -1,3 +1,7 @@
+# adbi (development version)
+
+- Requires R >= 4.1.0.
+
 # adbi 0.1.2 (2024-09-03)
 
 - Docs only update